agent_tool/pyproject.toml
[project]
name = "agent-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "duckdb>=1.3.0",
    "langchain>=0.3.25",
    "langchain-openai>=0.3.19",
    "langgraph>=0.4.8",
    "openai>=1.83.0",
    "python-dotenv>=1.1.0",
]
